The spatial index now contains the tuple byte positions

+++ src/main/java/org/bboxdb/storage/sstable/spatialindex/SpatialIndexEntry.java
@@ -45,7 +45,7 @@ public class SpatialIndexEntry implements BoundingBoxEntity {
 	 * Get the value
 	 * @return
 	 */
-	public long getValue() {
+	public int getValue() {
 		return value;
 	}
